refactor: Phase 3.5 — move 6 core/admin routers to domain modules (#512)
Move auth, roles, workspace → modules/core/
Move backup, legal, github_webhook → modules/admin/

Import replacements (7 total):
- async_audit_logger → audit_service (from modules.monitoring.service)
- async_session_manager → user_session_service (from modules.core.service)
- async_user_manager → user_service (from modules.core.service)
- async_role_manager → role_service (from modules.core.service)
- async_workspace_manager → workspace_service (from modules.core.service)
- async_kanban_project_manager → kanban_project_service (from modules.kanban.service)

Original app/routers/ files become thin facade re-exports.
backup.py had 0 db.integration imports — pure move.
legal.py and github_webhook.py had 1 inline import each.
